Save System Working

+++ Assets/ScriptalObjects/LevelData/LevelData.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "NewLevelData", menuName = "Level Data")]
@@ -8,7 +5,7 @@ public class LevelData : ScriptableObject
 {
     public int level;
 
-    public float GetLevel()
+    public int GetLevel()
     {
         return level;
     }
@@ -18,5 +15,3 @@ public class LevelData : ScriptableObject
         level = num;
     }
 }
-
-
